Add signup with email, refactor code, and more
