+++ src/pages/SignupPage.tsx
@@ -1,4 +1,5 @@
 import SignupBox from '@/components/SignupBox';
+import FadeIn from '@/components/ui/FadeIn';
 import GettingStartedNav from '@/components/ui/GettingStartedNav';
 import { motion, Variants } from 'framer-motion';
 
@@ -18,17 +19,14 @@ const pageVariants: Variants = {
 function SignupPage(): JSX.Element {
   return (
     <div className="overflow-x-hidden font-raleway bg-lighter dark:bg-purple-shade-400">
-      <motion.div
-        className="min-h-[101vh] bg-lighter dark:bg-purple-shade-400 pl-2 pr-2 sm:pl-10 sm:pr-10 relative flex flex-col"
-        variants={pageVariants}
-        animate="animate"
-        initial="initial"
-      >
-        <GettingStartedNav />
-        <div className="w-full flex sm:items-center sm:justify-center flex-1">
-          <SignupBox />
+      <FadeIn>
+        <div className="min-h-[101vh] bg-lighter dark:bg-purple-shade-400 pl-2 pr-2 sm:pl-10 sm:pr-10 relative flex flex-col">
+          <GettingStartedNav />
+          <div className="w-full flex sm:items-center sm:justify-center flex-1">
+            <SignupBox />
+          </div>
         </div>
-      </motion.div>
+      </FadeIn>
     </div>
   );
 }
